Implement form validation for monthly entries and adjust styling
Values entered in the NewPointModal while in TableView are now being validated.
The GCM name cannot be an empty string, and the monthly values must be valid
numbers. Additionally, if precipitation values are shown, the entered value
cannot be a negative number.
The TextField css is also adjusted to compensate for the potential error
messages under TextFields.