Move the loading indicator into the content card
The loading indicator now doesn't just hover above the entire page, but rather
it is a part of a transparent overlay located over the content card, giving a
more natural feedback to the end user that data is being fetched. Related
components have also been adjusted to support this change.